// apps/pdf-chat/lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      conversations: {
        Row: {
          account_id: string;
          created_at: string;
          document_id: string;
          id: number;
          name: string;
          reference_id: string;
        };
        Insert: {
          account_id: string;
          created_at?: string;
          document_id: string;
          id?: number;
          name: string;
          reference_id: string;
        };
        Update: {
          account_id?: string;
          created_at?: string;
          document_id?: string;
          id?: number;
          name?: string;
          reference_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'conversations_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'conversations_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'conversations_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'conversations_document_id_fkey';
            columns: ['document_id'];
            isOneToOne: false;
            referencedRelation: 'documents';
            referencedColumns: ['id'];
          },
        ];
      };
      credits_usage: {
        Row: {
          account_id: string;
          id: number;
          tokens_quota: number;
        };
        Insert: {
          account_id: string;
          id?: number;
          tokens_quota?: number;
        };
        Update: {
          account_id?: string;
          id?: number;
          tokens_quota?: number;
        };
        Relationships: [
          {
            foreignKeyName: 'credits_usage_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'credits_usage_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'credits_usage_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
        ];
      };
      documents: {
        Row: {
          account_id: string;
          content: string;
          created_at: string;
          id: string;
          title: string;
        };
        Insert: {
          account_id: string;
          content: string;
          created_at?: string;
          id?: string;
          title: string;
        };
        Update: {
          account_id?: string;
          content?: string;
          created_at?: string;
          id?: string;
          title?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'documents_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'documents_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'documents_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      messages: {
        Row: {
          account_id: string;
          conversation_id: number;
          created_at: string;
          id: number;
          sender: Database['public']['Enums']['sender'];
          text: string;
        };
        Insert: {
          account_id: string;
          conversation_id: number;
          created_at?: string;
          id?: number;
          sender: Database['public']['Enums']['sender'];
          text: string;
        };
        Update: {
          account_id?: string;
          conversation_id?: number;
          created_at?: string;
          id?: number;
          sender?: Database['public']['Enums']['sender'];
          text?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'messages_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'messages_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'messages_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'messages_conversation_id_fkey';
            columns: ['conversation_id'];
            isOneToOne: false;
            referencedRelation: 'conversations';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
// ...
  };
// ...
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema['Enums']
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
    ? PublicSchema['Enums'][PublicEnumNameOrOptions]
    : never;